Use GraphQL query for small multiples charts
